requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
motor==3.4.0
pymongo==4.7.2
httpx==0.27.0